Pipe: enhance the idempotence of start / stop pipe  (#11713)
As title. This commit is a follow-up to #11690, from now onwards, no-op start pipe and stop pipe will directly skip the entire procedure execution, avoiding the impact of rollback and other exceptional situations. At the same time, the procedure result is set to facilitate differentiation on CN side for identifying no-op start pipe and stop pipe.

It is important to note that when starting a RUNNING pipe or stopping a STOPPED pipe, we need to consider the `isStoppedByRuntimeException` status of the pipe. Otherwise, certain procedures might be mistakenly skipped.