[IOTDB-6287] Load: Add Progress Index for the Second Phase of Load TsFile (#11851)
Problems:
The Load TsFile process uses a two-phase transaction commit to ensure consistency. Currently, commands are sent directly to each DataNode through the Thrift framework during the second stage of validation, which does not ensure that the loaded TsFiles are organized within the DataRegionGroup. This does not ensure that the loaded TsFiles are organized within the DataRegionGroup, nor can they be correctly identified as progress by the Pipe system.

Solution:
A RecoverProgressIndex will be added to the second stage command, so that when a local load or a cross-DataNode load is performed, the ProgressIndex will be incorporated into the TsFileResource to complete the progress identification.

---------

Co-authored-by: Steve Yurong Su <rong@apache.org>