UDF: add an udf function for envelope demodulation analysis (#12146)
This commit adds a UDF function for envelope demodulation analysis. The function of envelope demodulation analysis is designed to realize the demodulation and envelope extraction of the signal by inputting one-dimensional floating-point numbers and the modulation frequency specified by the user. The goal of demodulation is to extract parts of interest from complex signals and make them easier to understand. For example, demodulation can find the envelope of the signal, that is, the trend of amplitude change. The realization principle is to convert the time domain signal into the frequency domain signal, and the algorithms involved include Hilbert transform (Hilbert), fast Fourier transform (FFT), fast Fourier inverse transform (IFFT) and so on.